Refactor: Move StatusError class to shared file
